Source -> Clean Up...
svn path=/trunk/javanlp/; revision=21717

+++ src/mt/base/TrieIntegerArrayIndex.java
@@ -32,7 +32,7 @@ public class TrieIntegerArrayIndex implements IntegerArrayIndex {
   }
 
   private long getTransition(int curState, int input) {
-    return ((long)curState << 32) | (long)input;
+    return ((long)curState << 32) | input;
   }
 
   public int size() { return lastAcceptingStateIdx; }
